Move base type to metadata conversion to own module
